# api_spec.yaml
openapi: 3.0.0
info:
  version: 1.1.0
  title: 'SET Web API'
  description: 'Web API for interacting with SET backend'
servers:
  - url: https://server.set.tp.vps.tecnico.ulisboa.pt
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
  schemas:
    Error:
      description: 'Error response body'
      properties:
        error:
          type: string
    Activity:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        date:
          type: string
          format: date-time
        duration:
          type: integer
        location:
          type: string
        type:
          type: string
          enum: [SpeedInterview, Workshop, Lecture]
        studentEnrolled:
          type: string
          enum: [Enrolled, Confirmed]
          description: "Note: Only returned if it's an authenticated student and it's enrolled or confirmed"
    ActivityManagementDetails:
      type: object
      properties:
        enrolledStudents:
          type: array
          items:
            $ref: '#/components/schemas/StudentPublic'
        confirmedStudents:
          type: array
          items:
            $ref: '#/components/schemas/StudentPublic'
    ActivityManagement:
      allOf:
        - $ref: '#/components/schemas/Activity'
        - $ref: '#/components/schemas/ActivityManagementDetails'
...
    User:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        name:
          type: string
        role:
          type: string
          enum: [Company, Student, Staff]
        image:
          type: string
    StudentPublicProperties:
      type: object
      properties:
        studentId:
          type: integer
        university:
          type: string
        course:
          type: string
        email:
          type: string
          format: email
          description: Only returned if requested by a company
        cvId:
          type: string
          format: uuid
          description: Only returned if requested by a company with PLATINUM category
...
    StudentPublic:
      allOf:
        - $ref: '#/components/schemas/User'
        - $ref: '#/components/schemas/StudentPublicProperties'
...
security:
  - cookieAuth: []
paths:
...
  /activities/{id}/enroll:
    post:
      description: Enroll student in an activity (Speed Interview / Workshop)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 'Successfully enrolled / Already enrolled'
        '400':
          description: 'Invalid activity id / Activity is not enrollable'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 'User not logged in'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 'No Activity found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /activities/{id}/cancel:
    post:
      description: Remove student from an activity (Speed Interview / Workshop)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 'Successfully removed'
        '400':
          description: 'Invalid activity id / Activity is not enrollable / Student not enrolled'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 'User not logged in'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 'No Activity found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /activities/manage/{id}:
    get:
      description: Returns the activity management details (STAFF route)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully return Activity management
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityManagement'
        '400':
          description: 'Invalid activity id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 'User not logged in'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 'User does not have the required privilege'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 'No Activity found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      description: Manage students in an activity (STAFF route)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  enum: [ENROLL, CONFIRM, DISCARD]
                userId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Successfully updated activity management details
        '400':
          description: 'Invalid activity id / Invalid request body / Student has to be enrolled to be confirmed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 'User not logged in'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 'User does not have the required privilege'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 'No Activity found / No StudentDetails found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...